+++ Items/Projectiles/RealityLaser.cs
@@ -30,7 +30,7 @@ namespace AAMod.Items.Projectiles
 
         public override void AI()
         {
-        	int num103 = (int)Player.FindClosest(projectile.Center, 1, 1);
+        	int num103 = Player.FindClosest(projectile.Center, 1, 1);
 			projectile.ai[1] += 1f;
 			if (projectile.ai[1] < 110f && projectile.ai[1] > 30f)
 			{
@@ -49,7 +49,7 @@ namespace AAMod.Items.Projectiles
 					projectile.velocity *= 1.02f;
 				}
 			}
-			projectile.rotation = (float)Math.Atan2((double)projectile.velocity.Y, (double)projectile.velocity.X) + 1.57f;
+			projectile.rotation = (float)Math.Atan2(projectile.velocity.Y, projectile.velocity.X) + 1.57f;
         	Lighting.AddLight(projectile.Center, ((255 - projectile.alpha) * 0.5f) / 255f, ((255 - projectile.alpha) * 0.05f) / 255f, ((255 - projectile.alpha) * 0.05f) / 255f);
         }
     }
